rename file containing BinaryPolynomial
