+++ src/openfermion/ops/_binary_polynomial.py
